increase search client request timeouts
